pyproject.toml
[project]
name = "ad-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "openai>=1.97.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]
